@@ -3,10 +3,10 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
-use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Foundation\Validation\ValidatesRequests;
 use Illuminate\Routing\Controller as BaseController;
 
-class Controller extends BaseController {
-    use AuthorizesRequests, DispatchesJobs, ValidatesRequests;
+class Controller extends BaseController
+{
+    use AuthorizesRequests, ValidatesRequests;
 }
