@@ -4,7 +4,8 @@ namespace App\Http\Middleware;
 
 use Illuminate\Cookie\Middleware\EncryptCookies as Middleware;
 
-class EncryptCookies extends Middleware {
+class EncryptCookies extends Middleware
+{
     /**
      * The names of the cookies that should not be encrypted.
      *
