@@ -4,7 +4,8 @@ namespace App\Http\Middleware;
 
 use Illuminate\Foundation\Http\Middleware\PreventRequestsDuringMaintenance as Middleware;
 
-class PreventRequestsDuringMaintenance extends Middleware {
+class PreventRequestsDuringMaintenance extends Middleware
+{
     /**
      * The URIs that should be reachable while maintenance mode is enabled.
      *
