@@ -4,7 +4,8 @@ namespace App\Http\Middleware;
 
 use Illuminate\Foundation\Http\Middleware\TrimStrings as Middleware;
 
-class TrimStrings extends Middleware {
+class TrimStrings extends Middleware
+{
     /**
      * The names of the attributes that should not be trimmed.
      *
