@@ -4,13 +4,15 @@ namespace App\Http\Middleware;
 
 use Illuminate\Http\Middleware\TrustHosts as Middleware;
 
-class TrustHosts extends Middleware {
+class TrustHosts extends Middleware
+{
     /**
      * Get the host patterns that should be trusted.
      *
      * @return array<int, string|null>
      */
-    public function hosts() {
+    public function hosts(): array
+    {
         return [
             $this->allSubdomainsOfApplicationUrl(),
         ];
