@@ -5,7 +5,8 @@ namespace App\Http\Middleware;
 use Illuminate\Http\Middleware\TrustProxies as Middleware;
 use Illuminate\Http\Request;
 
-class TrustProxies extends Middleware {
+class TrustProxies extends Middleware
+{
     /**
      * The trusted proxies for this application.
      *
