@@ -4,7 +4,8 @@ namespace App\Http\Middleware;
 
 use Illuminate\Foundation\Http\Middleware\VerifyCsrfToken as Middleware;
 
-class VerifyCsrfToken extends Middleware {
+class VerifyCsrfToken extends Middleware
+{
     /**
      * The URIs that should be excluded from CSRF verification.
      *
