@@ -4,7 +4,8 @@ namespace App\Providers;
 
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
 
-class AuthServiceProvider extends ServiceProvider {
+class AuthServiceProvider extends ServiceProvider
+{
     /**
      * The model to policy mappings for the application.
      *
@@ -16,12 +17,9 @@ class AuthServiceProvider extends ServiceProvider {
 
     /**
      * Register any authentication / authorization services.
-     *
-     * @return void
      */
-    public function boot() {
-        $this->registerPolicies();
-
+    public function boot(): void
+    {
         //
     }
 }
