@@ -5,13 +5,13 @@ namespace App\Providers;
 use Illuminate\Support\Facades\Broadcast;
 use Illuminate\Support\ServiceProvider;
 
-class BroadcastServiceProvider extends ServiceProvider {
+class BroadcastServiceProvider extends ServiceProvider
+{
     /**
      * Bootstrap any application services.
-     *
-     * @return void
      */
-    public function boot() {
+    public function boot(): void
+    {
         Broadcast::routes();
 
         require base_path('routes/channels.php');
