@@ -7,7 +7,8 @@ use Illuminate\Auth\Listeners\SendEmailVerificationNotification;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
 use Illuminate\Support\Facades\Event;
 
-class EventServiceProvider extends ServiceProvider {
+class EventServiceProvider extends ServiceProvider
+{
     /**
      * The event to listener mappings for the application.
      *
@@ -21,19 +22,17 @@ class EventServiceProvider extends ServiceProvider {
 
     /**
      * Register any events for your application.
-     *
-     * @return void
      */
-    public function boot() {
+    public function boot(): void
+    {
         //
     }
 
     /**
      * Determine if events and listeners should be automatically discovered.
-     *
-     * @return bool
      */
-    public function shouldDiscoverEvents() {
+    public function shouldDiscoverEvents(): bool
+    {
         return false;
     }
 }
