@@ -8,7 +8,8 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\RateLimiter;
 use Illuminate\Support\Facades\Route;
 
-class RouteServiceProvider extends ServiceProvider {
+class RouteServiceProvider extends ServiceProvider
+{
     /**
      * The path to the "home" route for your application.
      *
@@ -20,10 +21,9 @@ class RouteServiceProvider extends ServiceProvider {
 
     /**
      * Define your route model bindings, pattern filters, and other route configuration.
-     *
-     * @return void
      */
-    public function boot() {
+    public function boot(): void
+    {
         $this->configureRateLimiting();
 
         $this->routes(function () {
@@ -38,10 +38,9 @@ class RouteServiceProvider extends ServiceProvider {
 
     /**
      * Configure the rate limiters for the application.
-     *
-     * @return void
      */
-    protected function configureRateLimiting() {
+    protected function configureRateLimiting(): void
+    {
         RateLimiter::for('api', function (Request $request) {
             return Limit::perMinute(60)->by($request->user()?->id ?: $request->ip());
         });
